requirements.txt
pandas
numpy
streamlit
websockets
python-dotenv
jupyterlab
scipy
plotly
shapely